requirements.txt
matplotlib==3.3.2
pandas==1.1.3
Pillow==9.3.0
PyFMI==2.9.5
PyYAML==6.0
requests==2.24.0
streamlit==1.3.1
streamlit_authenticator==0.2.1
streamlit_lottie==0.0.3

